requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
Django==4.2.15
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.31.0
whitenoise==6.7.0
